test(multi.test.js): increased testing coverage of multi module
